Move setImmediate to rts.setimmediate.mjs
